+++ wifipumpkin3/extensions/ap.py
@@ -48,6 +48,7 @@ class Ap(ExtensionUI):
                 "ssid",
                 "channel",
                 "interface",
+                "interface_net",
                 "status",
                 "security",
                 "hostapd_config",
@@ -62,6 +63,9 @@ class Ap(ExtensionUI):
                 self.root.conf.get("accesspoint", self.root.commands["ssid"]),
                 self.root.conf.get("accesspoint", self.root.commands["channel"]),
                 self.root.conf.get("accesspoint", self.root.commands["interface"]),
+                setcolor("default", color="green") 
+                if not self.root.conf.get("accesspoint", self.root.commands["interface_net"]) else
+                self.root.conf.get("accesspoint", self.root.commands["interface_net"]  ),
                 setcolor("is Running", color="green")
                 if status_ap
                 else setcolor("not Running", color="red"),